@@ -15,7 +15,7 @@ import (
 // might have some errors.
 
 // LoadConfigAsMap loads the config.json file as map[string]interface{}
-func LoadConfigAsMap() (map[string]interface{}, error) {
+func LoadConfigAsMap() (map[string]any, error) {
 	err := CheckSuspiciousLocation()
 	if err != nil {
 		return nil, burrito.PassError(err)
@@ -27,7 +27,7 @@ func LoadConfigAsMap() (map[string]interface{}, error) {
 				"Please make sure to run this command in a Regolith project directory.\n" +
 				"If you want to create new Regolith project here, use \"regolith init\".")
 	}
-	var configJson map[string]interface{}
+	var configJson map[string]any
 	err = jsonc.Unmarshal(file, &configJson)
 	if err != nil {
 		return nil, burrito.WrapErrorf(err, jsonUnmarshalError, ConfigFilePath)
@@ -37,14 +37,14 @@ func LoadConfigAsMap() (map[string]interface{}, error) {
 
 // dataPathFromConfigMap returns the value of the data path from the config
 // file map, without parsing it to a Config object.
-func dataPathFromConfigMap(config map[string]interface{}) (string, error) {
+func dataPathFromConfigMap(config map[string]any) (string, error) {
 	return FindByJSONPath[string](config, "regolith/dataPath")
 }
 
 // filterDefinitionFromConfigMap returns the filter definitions as map from
 // the config file map, without parsing it to a Config object.
 func filterDefinitionsFromConfigMap(
-	config map[string]interface{},
-) (map[string]interface{}, error) {
-	return FindByJSONPath[map[string]interface{}](config, "regolith/filterDefinitions")
+	config map[string]any,
+) (map[string]any, error) {
+	return FindByJSONPath[map[string]any](config, "regolith/filterDefinitions")
 }
