@@ -38,16 +38,20 @@ func EvalString(expression string, ctx RunContext) (string, error) {
 	return "", burrito.WrapErrorf(err, "Expression evaluated to non-string value: %s", expression)
 }
 
-func prepareScope(ctx RunContext) map[string]interface{} {
+func prepareScope(ctx RunContext) map[string]any {
 	semverString, err := utils.ParseSemverString(Version)
 	if err != nil {
 		semverString = utils.Semver{}
 	}
-	projectData := map[string]interface{}{
+	projectData := map[string]any{
 		"name":   ctx.Config.Name,
 		"author": ctx.Config.Author,
 	}
-	return map[string]interface{}{
+	mode := "run"
+	if ctx.IsInWatchMode() {
+		mode = "watch"
+	}
+	return map[string]any{
 		"os":             runtime.GOOS,
 		"arch":           runtime.GOARCH,
 		"debug":          burrito.PrintStackTrace,
@@ -56,5 +60,7 @@ func prepareScope(ctx RunContext) map[string]interface{} {
 		"filterLocation": ctx.AbsoluteLocation,
 		"settings":       ctx.Settings,
 		"project":        projectData,
+		"mode":           mode,
+		"initial":        ctx.Initial,
 	}
 }
