@@ -2,6 +2,7 @@ package regolith
 
 import (
 	"encoding/json"
+	"io/fs"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -25,7 +26,7 @@ type NimFilter struct {
 }
 
 func NimFilterDefinitionFromObject(
-	id string, obj map[string]interface{},
+	id string, obj map[string]any,
 ) (*NimFilterDefinition, error) {
 	filter := &NimFilterDefinition{FilterDefinition: *FilterDefinitionFromObject(id)}
 	scriptObj, ok := obj["script"]
@@ -96,8 +97,8 @@ func (f *NimFilter) Run(context RunContext) (bool, error) {
 	return context.IsInterrupted(), nil
 }
 
-func (f *NimFilterDefinition) CreateFilterRunner(runConfiguration map[string]interface{}) (FilterRunner, error) {
-	basicFilter, err := filterFromObject(runConfiguration)
+func (f *NimFilterDefinition) CreateFilterRunner(runConfiguration map[string]any, id string) (FilterRunner, error) {
+	basicFilter, err := filterFromObject(runConfiguration, id)
 	if err != nil {
 		return nil, burrito.WrapError(err, filterFromObjectError)
 	}
@@ -173,8 +174,8 @@ func (f *NimFilter) Check(context RunContext) error {
 
 func hasNimble(filterPath string) bool {
 	nimble := false
-	filepath.Walk(filterPath, func(path string, info os.FileInfo, err error) error {
-		if err != nil || info.IsDir() {
+	filepath.WalkDir(filterPath, func(path string, d fs.DirEntry, err error) error {
+		if err != nil || d.IsDir() {
 			return nil
 		}
 		if filepath.Ext(path) == ".nimble" {
