@@ -11,6 +11,7 @@ import (
 	"path/filepath"
 	"reflect"
 	"runtime"
+	"slices"
 	"strconv"
 	"strings"
 	"time"
@@ -32,6 +33,17 @@ const appDataResolverCachePath = "regolith/resolver-cache"
 const appDataFilterCachePath = "regolith/filter-cache"
 
 var Version = "unversioned"
+
+// ComMojangPathType is used to specify the type of the com.mojang path you
+// need in some functions. Since Minecraft 1.21.120, there are two separate,
+// paths: shared (recommended for keeping packs) and user path (used for
+// storing Minecraft worlds).
+type ComMojangPathType int
+
+const (
+	WorldPath ComMojangPathType = iota
+	PacksPath
+)
 
 // nth returns the ordinal numeral of the index of a table. For example:
 // nth(0) returns "1st", nth(1) returns "2nd", etc.
@@ -136,7 +148,7 @@ func RunGitProcess(args []string, workingDir string) ([]string, error) {
 	out, _ := cmd.StdoutPipe()
 	err, _ := cmd.StderrPipe()
 	completeOutput := make([]string, 0)
-	logFunc := func(template string, args ...interface{}) {
+	logFunc := func(template string, args ...any) {
 		completeOutput = append(completeOutput, fmt.Sprintf(template, args...))
 	}
 	go LogStd(out, logFunc, "git")
@@ -146,7 +158,7 @@ func RunGitProcess(args []string, workingDir string) ([]string, error) {
 }
 
 // LogStd logs the output of a sub-process
-func LogStd(in io.ReadCloser, logFunc func(template string, args ...interface{}), outputLabel string) {
+func LogStd(in io.ReadCloser, logFunc func(template string, args ...any), outputLabel string) {
 	scanner := bufio.NewScanner(in)
 	for scanner.Scan() {
 		logFunc("[%s] %s", outputLabel, scanner.Text())
@@ -250,43 +262,59 @@ func acquireSessionLock(dotRegolithPath string) (func() error, error) {
 	return unlockFunc, nil
 }
 
-func splitPath(path string) []string {
-	parts := make([]string, 0)
-	for true {
-		part := ""
-		path, part = filepath.Split(path)
-		if strings.HasSuffix(path, "/") || strings.HasSuffix(path, "\\") {
-			path = path[0 : len(path)-1]
-		}
-		if path == "" && part != "" {
-			parts = append([]string{part}, parts...)
-			break
-		}
-		if part == "" || path == "" {
-			break
-		}
-		parts = append([]string{part}, parts...)
-	}
-	return parts
-}
-
 func ResolvePath(path string) (string, error) {
-	// Resolve the path
-	parts := splitPath(path)
-	for i, part := range parts {
-		if strings.HasPrefix(part, "%") && strings.HasSuffix(part, "%") {
-			envVar := part[1 : len(part)-1]
-			envVarValue, exists := os.LookupEnv(envVar)
-			if !exists {
-				return "", burrito.WrapErrorf(
-					os.ErrNotExist,
-					"Environment variable %s does not exist.",
-					envVar)
-			}
-			parts[i] = envVarValue
+	// Expand %VAR% style markers
+	parts := make([]string, 0)
+	parsed := 0
+	for {
+		split := strings.Index(path[parsed:], "%")
+		if split == -1 {
+			// End - Append the remaining path
+			parts = append(parts, path[parsed:])
+			break
+		}
+		// Found split location
+		parts = append(parts, path[parsed:parsed+split])
+		parsed += split + 1
+
+		// Bounds check
+		if parsed >= len(path) {
+			parts = append(parts, "")
+			break
 		}
 	}
-	return filepath.Clean(filepath.Join(parts...)), nil
+
+	iterations := len(parts)
+	if iterations%2 == 0 {
+		// If number of iterations is even, that means that the number of %
+		// markers is odd, therefore the last part is not a variable name
+		// because it lacks the trailing % marker
+		iterations -= 1
+		// Readd the % that we just removed from the last part
+		lastPart := parts[len(parts)-1]
+		parts[len(parts)-1] = "%" + lastPart
+	}
+	// Every even part is a variable name that we can skip
+	for i := 1; i < iterations; i += 2 {
+		envVar := parts[i]
+		envVarValue, exists := os.LookupEnv(envVar)
+		if !exists {
+			return "", burrito.WrapErrorf(
+				os.ErrNotExist,
+				"Environment variable %s does not exist.",
+				envVar)
+		}
+		parts[i] = envVarValue
+	}
+	path = filepath.Join(parts...)
+
+	// Expand $VAR and ${VAR} markers
+	path = os.Expand(path, func(v string) string {
+		val, _ := os.LookupEnv(v)
+		return val
+	})
+
+	return filepath.Clean(path), nil
 }
 
 type measure struct {
@@ -301,7 +329,7 @@ type measure struct {
 var lastMeasure *measure
 var EnableTimings = false
 
-func MeasureStart(name string, args ...interface{}) {
+func MeasureStart(name string, args ...any) {
 	if !EnableTimings {
 		return
 	}
@@ -329,16 +357,11 @@ func MeasureEnd() {
 }
 
 func stringInSlice(a string, list []string) bool {
-	for _, b := range list {
-		if b == a {
-			return true
-		}
-	}
-	return false
+	return slices.Contains(list, a)
 }
 
 // FindByJSONPath finds a value in a JSON element by a simple path. Returns nil and an error if the path is not found or invalid.
-func FindByJSONPath[T any](obj interface{}, path string) (T, error) {
+func FindByJSONPath[T any](obj any, path string) (T, error) {
 	var empty T
 	if obj == nil {
 		return empty, burrito.WrappedErrorf("Object is empty")
@@ -356,14 +379,14 @@ func FindByJSONPath[T any](obj interface{}, path string) (T, error) {
 			continue
 		}
 		currentPath += part + "->"
-		if m, ok := value.(map[string]interface{}); ok {
+		if m, ok := value.(map[string]any); ok {
 			value = m[part]
 			if value == nil {
 				return empty, burrito.WrappedErrorf(jsonPathMissingError, currentPath[:len(currentPath)-2])
 			}
 			continue
 		}
-		if a, ok := value.([]interface{}); ok {
+		if a, ok := value.([]any); ok {
 			index, err := strconv.Atoi(part)
 			if err != nil {
 				return empty, burrito.WrapErrorf(err, "Invalid index %s at %s", part, currentPath[:len(currentPath)-2])
@@ -432,11 +455,61 @@ func EscapePathPart(s string) string {
 }
 
 // SliceAny returns true if any of the elements in the slice satisfy the predicate.
-func SliceAny[T interface{}](slice []T, predicate func(T) bool) bool {
-	for _, item := range slice {
-		if predicate(item) {
-			return true
-		}
+func SliceAny[T any](slice []T, predicate func(T) bool) bool {
+	return slices.ContainsFunc(slice, predicate)
+}
+
+func isSymlinkTo(path, target string) bool {
+	info, err := os.Lstat(path)
+	if err != nil || info.Mode()&os.ModeSymlink == 0 {
+		return false
 	}
-	return false
+	dest, err := os.Readlink(path)
+	if err != nil {
+		return false
+	}
+	if !filepath.IsAbs(dest) {
+		dest = filepath.Join(filepath.Dir(path), dest)
+	}
+	absDest, _ := filepath.Abs(dest)
+	absTarget, _ := filepath.Abs(target)
+	return absDest == absTarget
+}
+
+func isSymlink(path string) bool {
+	info, err := os.Lstat(path)
+	if err != nil || info.Mode()&os.ModeSymlink == 0 {
+		return false
+	}
+	return true
+}
+
+func createDirLink(link, target string) error {
+	if _, err := os.Lstat(link); err == nil {
+		return burrito.WrappedErrorf(
+			"Failed to create symlink, path already exists.\n"+
+				"Link: %s\n"+
+				"Target: %s",
+			link, target)
+	}
+	linkDir := filepath.Dir(link)
+	if err := os.MkdirAll(linkDir, 0755); err != nil {
+		return burrito.WrapErrorf(err, osMkdirError, linkDir)
+	}
+	if err := os.MkdirAll(target, 0755); err != nil {
+		return burrito.WrapErrorf(err, osMkdirError, target)
+	}
+	absTarget, err := filepath.Abs(target)
+	if err != nil {
+		return burrito.WrapErrorf(err, filepathAbsError, target)
+	}
+	absLink, err := filepath.Abs(link)
+	if err != nil {
+		return burrito.WrapErrorf(err, filepathAbsError, link)
+	}
+	err = os.Symlink(absTarget, absLink)
+	if err != nil {
+		return burrito.PassError(err)
+	}
+	return nil
 }
