@@ -1,10 +1,15 @@
 package regolith
 
+import "slices"
+
 type Experiment int
 
 const (
 	// SizeTimeCheck is an experiment that checks the size and modification time when exporting
 	SizeTimeCheck Experiment = iota
+	// SymlinkExport links the temporary build directory with the export
+	// target using hard links when possible.
+	SymlinkExport
 )
 
 // The descriptions shouldn't be too wide, the text with their description is
@@ -16,6 +21,10 @@ the file has changed. This experiment applies to 'run' and 'watch'
 commands.
 `
 
+const symlinkExportDesc = `
+Creates links from the tmp directory to the export target so that files
+written to tmp are immediately reflected in the export location.`
+
 type ExperimentInfo struct {
 	Name        string
 	Description string
@@ -23,6 +32,7 @@ type ExperimentInfo struct {
 
 var AvailableExperiments = map[Experiment]ExperimentInfo{
 	SizeTimeCheck: {"size_time_check", sizeTimeCheckDesc},
+	SymlinkExport: {"symlink_export", symlinkExportDesc},
 }
 
 var EnabledExperiments []string
@@ -31,10 +41,5 @@ func IsExperimentEnabled(exp Experiment) bool {
 	if EnabledExperiments == nil {
 		return false
 	}
-	for _, e := range EnabledExperiments {
-		if e == AvailableExperiments[exp].Name {
-			return true
-		}
-	}
-	return false
+	return slices.Contains(EnabledExperiments, AvailableExperiments[exp].Name)
 }
