@@ -122,7 +122,8 @@ func listFiles(path string) ([]string, error) {
 				if err != nil {
 					return burrito.WrapErrorf(err, osRelError, path, s)
 				}
-				result = append(result, relpath)
+				normalizedRelPath := strings.ReplaceAll(relpath, "\\", "/")
+				result = append(result, normalizedRelPath)
 			}
 			return nil
 		})
@@ -133,12 +134,14 @@ func listFiles(path string) ([]string, error) {
 }
 
 // checkDeletionSafety checks whether it's safe to delete files from given path
-// based on the list of removable files. The removableFiles list must be
-// sorted. The function relies on filepath.WalkDir walking files
-// alphabetically. It returns nil value when it's safe to delete the files or
-// an error in opposite case.
+// based on the list of removable files. It returns nil value when it's safe to
+// delete the files or an error in opposite case.
 func checkDeletionSafety(path string, removableFiles []string) error {
-	i := 0 // current index on the removableFiles list to check
+	removableSet := make(map[string]struct{}, len(removableFiles))
+	for _, f := range removableFiles {
+		normalized := strings.ReplaceAll(f, "\\", "/")
+		removableSet[normalized] = struct{}{}
+	}
 	stats, err := os.Stat(path)
 	if err != nil {
 		if os.IsNotExist(err) {
@@ -163,18 +166,9 @@ func checkDeletionSafety(path string, removableFiles []string) error {
 			s = relpath // remove path from the file path
 			const notRegolithFileError = "File is not on the list of files" +
 				" created by Regolith.\nPath: %s"
-			for {
-				if i >= len(removableFiles) {
-					return burrito.WrappedErrorf(notRegolithFileError, s)
-				}
-				currPath := removableFiles[i]
-				i++
-				cmpVal := compareFilePaths(s, currPath)
-				if cmpVal == 0 { // found path on the list
-					break
-				} else if cmpVal < 0 { // this path won't be on the list
-					return burrito.WrappedErrorf(notRegolithFileError, s)
-				}
+			normalizedS := strings.ReplaceAll(s, "\\", "/")
+			if _, ok := removableSet[normalizedS]; !ok {
+				return burrito.WrappedErrorf(notRegolithFileError, s)
 			}
 			return nil
 		})
@@ -182,34 +176,4 @@ func checkDeletionSafety(path string, removableFiles []string) error {
 		return burrito.PassError(err)
 	}
 	return nil
-}
-
-// compareFilePaths compares two filepaths to oder them lexicographically.
-// This is not the same as comparing the file paths as strings because
-// "." < "/" and "." < "\\" but the "text.txt" should be greater than
-//  "text/text.txt" ("text.txt" > "text/text.txt"). This is the same order
-// that you would get when you use filepath.Walk.
-// The function returns -1 when "a" < "b", 0 when "a" == "b" and 1 when
-// "a" > "b".
-func compareFilePaths(a, b string) int {
-	a = strings.Replace(a, "\\", "/", -1)
-	b = strings.Replace(b, "\\", "/", -1)
-	aSlice := strings.Split(a, string("/"))
-	bSlice := strings.Split(b, string("/"))
-	for i := 0; i < len(aSlice) && i < len(bSlice); i++ {
-		if cmp := strings.Compare(aSlice[i], bSlice[i]); cmp != 0 {
-			return cmp
-		} // else - they're the same
-	}
-	if len(aSlice) < len(bSlice) {
-		// This shouldn't really happen because you can't use exactly the same
-		// name for file and directory.
-		return -1
-	}
-	if len(aSlice) > len(bSlice) {
-		// This shouldn't really happen because you can't use exactly the same
-		// name for file and directory.
-		return 1
-	}
-	return 0
 }
