@@ -20,7 +20,11 @@ func TestDevelopmentStandardExportLocation(t *testing.T) {
 		t.Skip("Skipping test on local machine")
 	}
 	_testCustomDevelopmentExportLocation(
-		t, regolith.FindStandardMojangDir, "standard",
+		t,
+		func() (string, error) {
+			return regolith.FindStandardMojangDir(regolith.PacksPath)
+		},
+		"standard",
 		"TestDevelopmentStandardExportLocation")
 }
 
@@ -46,7 +50,11 @@ func TestDevelopmentPreviewExportLocation(t *testing.T) {
 		t.Skip("Skipping test on local machine")
 	}
 	_testCustomDevelopmentExportLocation(
-		t, regolith.FindPreviewDir, "preview",
+		t,
+		func() (string, error) {
+			return regolith.FindPreviewDir(regolith.PacksPath)
+		},
+		"preview",
 		"TestDevelopmentPreviewExportLocation")
 }
 
